backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.21.6
filelock==3.12.2
gunicorn==20.1.0
identify==2.5.24
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
Markdown==3.4.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pre-commit==3.3.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.23.1
whitenoise==6.5.0
